refactored header and changed wallet connect buttons layout to menu
