+++ packages/app/src/components/Header/assets.ts
@@ -1,0 +1,11 @@
+export const logoUri = `data:image/svg+xml;utf8,<svg width="179" height="24" viewBox="0 0 179 24" fill="none" xmlns="http://www.w3.org/2000/svg"> <path d="M38.3916 14.2049C38.3881 10.6367 35.4518 7.74715 31.8282 7.74194C28.2064 7.74541 25.2701 10.6367 25.2666 14.2049C25.2701 17.773 28.2064 20.6626 31.8282 20.6678C35.4518 20.6644 38.3881 17.773 38.3916 14.2049ZM27.9473 14.2049C27.9473 12.0945 29.6868 10.3851 31.83 10.3833C33.9731 10.3833 35.7092 12.0945 35.7127 14.2049C35.7127 16.3152 33.9731 18.0247 31.83 18.0264C29.6868 18.0247 27.9508 16.3135 27.9473 14.2049Z" fill="#5B7AC6"/> <path d="M52.6888 14.2049C52.6853 10.6367 49.749 7.74715 46.1254 7.74194C42.5018 7.74541 39.5673 10.6367 39.562 14.2049C39.5655 17.773 42.5018 20.6644 46.1254 20.6678C49.749 20.6644 52.6835 17.773 52.6888 14.2049ZM42.2445 14.2049C42.2445 12.0945 43.984 10.3851 46.1272 10.3833C48.2703 10.3833 50.0063 12.0963 50.0099 14.2049C50.0063 16.3152 48.2703 18.0247 46.1272 18.0264C43.984 18.0247 42.248 16.3135 42.2445 14.2049Z" fill="#5B7AC6"/> <path d="M96.6605 14.2049C96.657 10.6367 93.7207 7.74715 90.0971 7.74194C86.4735 7.74541 83.539 10.6367 83.5337 14.2049C83.5372 17.773 86.4735 20.6644 90.0971 20.6678C93.7207 20.6644 96.6552 17.773 96.6605 14.2049ZM86.2162 14.2049C86.2162 12.0945 87.9557 10.3851 90.0988 10.3833C92.242 10.3833 93.978 12.0963 93.9815 14.2049C93.978 16.3152 92.242 18.0247 90.0988 18.0264C87.9557 18.0247 86.2197 16.3135 86.2162 14.2049Z" fill="#2B4483"/> <path d="M64.2537 8.99837C63.1293 8.18096 61.7686 7.74188 60.371 7.74362C56.7474 7.74709 53.8129 10.6384 53.8076 14.2048C53.8111 17.773 56.7474 20.6643 60.371 20.6678C63.9946 20.6643 66.9291 17.773 66.9326 14.2048C66.9326 14.0573 66.9274 13.9098 66.9185 13.7657H66.9326V3.74854H64.2519V8.99837H64.2537ZM60.371 18.0281C58.2279 18.0264 56.4918 16.3152 56.4901 14.2048C56.4936 12.0945 58.2296 10.385 60.371 10.3833C62.5142 10.3833 64.2502 12.0945 64.2537 14.2048C64.252 16.3152 62.5142 18.0246 60.371 18.0281Z" fill="#5B7AC6"/> <path d="M24.5477 3.33203H20.7566L19.5687 4.50175C18.3456 3.88739 16.9938 3.56806 15.6208 3.57153C10.8287 3.57673 6.94423 7.40174 6.93894 12.1205C6.93894 13.4707 7.26323 14.8018 7.88714 16.0045L4.4292 19.4061H8.22377L9.48745 18.1617C11.1124 19.7671 13.3208 20.6695 15.6226 20.6678C20.4165 20.6626 24.2992 16.8376 24.3044 12.117C24.3044 11.6745 24.2692 11.2337 24.2005 10.7963H16.962L14.2813 13.436H21.4722C20.843 16.1191 18.4196 18.0229 15.6226 18.0264C12.3091 18.0229 9.62493 15.378 9.6214 12.117C9.62493 8.85434 12.3091 6.2112 15.6226 6.20773C17.2141 6.20426 18.7421 6.8273 19.8648 7.938L20.8324 6.98696L24.5477 3.33203Z" fill="#5B7AC6"/> <path d="M75.7945 6.88129C74.9168 7.22145 74.1907 7.76639 73.5844 8.46579C73.2319 8.87363 72.9534 9.33353 72.7366 9.82988C72.571 10.2099 72.4511 10.6022 72.3736 11.0065C72.3489 11.135 72.3383 11.2669 72.3154 11.3953C72.2449 11.7927 72.2431 12.1936 72.2661 12.5928C72.2802 12.8288 72.326 13.0648 72.3683 13.2991C72.4176 13.5803 72.4846 13.8579 72.5868 14.1269C72.7102 14.4463 72.8389 14.7621 73.0433 15.0433C73.1508 15.1908 73.2301 15.3591 73.3359 15.5066C73.6478 15.9405 74.0338 16.3067 74.4585 16.6312C74.7317 16.8395 75.0313 17.0044 75.338 17.1588C75.7839 17.3827 76.258 17.525 76.7427 17.6343C76.963 17.6847 77.1974 17.676 77.4265 17.6812C77.7015 17.6881 77.9782 17.6812 78.2531 17.6829C78.5721 17.6864 78.8823 17.6222 79.1802 17.5267C79.5062 17.4226 79.827 17.2925 80.1354 17.1449C80.4121 17.0113 80.6712 16.843 80.9338 16.6833C81.1206 16.5705 81.3004 16.4473 81.4819 16.3293C81.4996 16.3171 81.5225 16.3119 81.5771 16.2876C81.8838 17.3046 82.1869 18.3008 82.4901 19.3074C82.4249 19.349 82.372 19.3854 82.3156 19.4184C81.8556 19.6787 81.3797 19.9044 80.888 20.1022C80.4756 20.2688 80.0491 20.3677 79.6243 20.4788C79.3476 20.5517 79.0533 20.5552 78.7678 20.602C78.1368 20.7062 77.5023 20.6593 76.8696 20.6263C76.6405 20.6142 76.4078 20.5899 76.1769 20.5361C75.6993 20.4233 75.2164 20.3157 74.7599 20.1439C73.8452 19.7985 73.008 19.316 72.252 18.6999C71.9523 18.4552 71.6774 18.1897 71.4095 17.912C71.027 17.5146 70.701 17.0755 70.4067 16.6121C70.3185 16.4716 70.211 16.3414 70.1352 16.1939C70.0066 15.9422 69.8938 15.6819 69.7775 15.4233C69.6946 15.2376 69.61 15.0537 69.5448 14.8628C69.4814 14.6823 69.4567 14.4879 69.388 14.3109C69.3227 14.146 69.3086 13.9725 69.2663 13.8059C69.2117 13.5907 69.1923 13.3651 69.1765 13.1429C69.15 12.7785 69.1289 12.4123 69.1201 12.0461C69.113 11.7406 69.1271 11.4335 69.1377 11.1263C69.1412 11.0256 69.1571 10.9267 69.1747 10.826C69.2452 10.3939 69.3421 9.96698 69.4743 9.55046C69.7175 8.79032 70.0824 8.08919 70.5247 7.42276C70.7609 7.06699 71.0182 6.7303 71.309 6.41965C71.4588 6.25825 71.621 6.109 71.7796 5.95454C71.8642 5.87298 71.9506 5.79314 72.0405 5.71678C72.1973 5.58315 72.3489 5.44084 72.5181 5.32283C72.8071 5.11978 73.0979 4.91325 73.4064 4.74491C73.8029 4.52798 74.2118 4.33013 74.6277 4.15832C74.9714 4.01601 75.3327 3.91014 75.6993 3.83031C76.1822 3.72618 76.6616 3.6099 77.1622 3.6047C77.4177 3.60123 77.6715 3.54049 77.9253 3.54743C78.3941 3.55784 78.8612 3.58214 79.3229 3.67585C79.6384 3.74007 79.9574 3.78172 80.2711 3.85114C80.4262 3.88585 80.5725 3.96221 80.7223 4.01774C81.1259 4.16873 81.5225 4.33187 81.9067 4.52798C82.1217 4.63905 82.3297 4.75879 82.557 4.89069C82.097 5.79488 81.66 6.70254 81.1788 7.60325C81.0836 7.53904 81.0114 7.48697 80.9356 7.44185C80.4985 7.17806 80.0473 6.94724 79.5468 6.81361C79.3705 6.76675 79.2031 6.68518 79.018 6.66262C78.9299 6.65221 78.8453 6.6175 78.7572 6.60882C78.4822 6.57758 78.2055 6.53419 77.9288 6.52725C77.1921 6.50816 76.4766 6.61923 75.7857 6.8865" fill="#2B4483"/> <path d="M122.295 13.49C122.373 12.6153 122.671 11.8118 123.124 11.0603C123.217 10.9059 123.309 10.7497 123.42 10.6074C123.58 10.4026 123.756 10.2117 123.926 10.0138C124.007 9.91839 124.081 9.81426 124.172 9.72922C124.319 9.59385 124.474 9.46543 124.631 9.34221C124.773 9.2294 124.923 9.12701 125.071 9.02288C125.464 8.75041 125.894 8.55082 126.344 8.39116C126.636 8.28703 126.927 8.18117 127.225 8.09266C127.377 8.04754 127.539 8.04059 127.697 8.01977C128.133 7.9625 128.57 7.91217 129.01 7.9382C129.174 7.94861 129.336 7.98159 129.499 8.00415C129.518 8.00762 129.537 8.00935 129.555 8.00935C129.983 7.99894 130.394 8.10828 130.797 8.22629C131.453 8.41893 132.058 8.72958 132.627 9.10097C132.711 9.15651 132.8 9.21031 132.882 9.26758C132.986 9.33873 132.995 9.37691 132.941 9.48799C132.655 10.0781 132.368 10.6664 132.084 11.2564C132.022 11.3831 131.975 11.5168 131.915 11.6435C131.888 11.699 131.851 11.7511 131.811 11.817C131.7 11.7458 131.601 11.6782 131.499 11.6209C131.129 11.4092 130.76 11.1974 130.385 10.9944C130.179 10.8833 129.95 10.8313 129.717 10.8017C129.259 10.7427 128.797 10.7115 128.335 10.767C127.71 10.8434 127.16 11.0725 126.67 11.4751C126.287 11.7892 125.972 12.1485 125.743 12.5789C125.579 12.8843 125.443 13.2054 125.396 13.5473C125.332 13.9881 125.297 14.4324 125.362 14.8819C125.41 15.2099 125.485 15.524 125.611 15.8329C125.808 16.3189 126.116 16.7215 126.506 17.0738C126.823 17.3601 127.179 17.5823 127.584 17.7159C128.141 17.8999 128.718 17.9051 129.299 17.853C129.52 17.8322 129.735 17.8044 129.948 17.7454C130.463 17.5996 130.928 17.3567 131.368 17.0599C131.603 16.9002 131.843 16.7475 132.095 16.5826C132.417 17.4747 132.733 18.3494 133.052 19.2362C132.505 19.5694 131.966 19.92 131.36 20.1525C131.129 20.241 130.9 20.3365 130.667 20.4267C130.628 20.4424 130.582 20.4424 130.542 20.4545C129.701 20.72 128.836 20.6853 127.972 20.6419C127.771 20.6315 127.565 20.6142 127.371 20.5638C127.036 20.4753 126.71 20.3591 126.381 20.2532C126.085 20.1577 125.811 20.0172 125.54 19.8696C124.347 19.2188 123.471 18.2747 122.866 17.079C122.637 16.6243 122.489 16.1436 122.373 15.6507C122.311 15.3886 122.311 15.1214 122.26 14.861C122.2 14.5539 122.221 14.2484 122.23 13.9412C122.234 13.8198 122.271 13.6983 122.294 13.5768C122.294 13.549 122.294 13.5213 122.295 13.4935V13.49Z" fill="#2B4483"/> <path d="M154.199 20.6661C154.152 20.5724 154.101 20.4786 154.056 20.3814C153.947 20.142 153.843 19.9007 153.734 19.663C153.581 19.3297 153.436 18.9913 153.265 18.6668C153.091 18.3353 152.953 17.9865 152.793 17.6498C152.641 17.3322 152.497 17.0146 152.343 16.6988C152.224 16.4506 152.091 16.2076 151.973 15.9577C151.862 15.7234 151.764 15.4822 151.654 15.2462C151.612 15.1542 151.549 15.0691 151.506 14.9772C151.427 14.8001 151.363 14.6144 151.282 14.4392C151.046 13.9254 150.803 13.4152 150.563 12.9032C150.329 12.4052 150.094 11.9088 149.862 11.4107C149.626 10.9074 149.393 10.4007 149.153 9.89913C148.963 9.49996 148.76 9.10427 148.572 8.70338C148.462 8.47082 148.367 8.22959 148.26 7.97447C148.366 7.97447 148.452 7.97447 148.536 7.97447C149.509 7.97447 150.484 7.97447 151.457 7.97447C151.51 7.97447 151.564 7.97621 151.617 7.97795C151.695 7.97795 151.73 8.02133 151.758 8.09249C151.882 8.3962 152.014 8.69644 152.139 8.99841C152.252 9.27262 152.354 9.55029 152.469 9.82276C152.585 10.097 152.715 10.3642 152.83 10.6402C153.011 11.074 153.181 11.5114 153.364 11.9435C153.542 12.3652 153.736 12.7818 153.915 13.2035C154.032 13.476 154.129 13.7571 154.245 14.0313C154.37 14.3298 154.509 14.6214 154.64 14.9181C154.733 15.1299 154.82 15.3433 154.924 15.588C155.986 13.0438 157.039 10.5256 158.089 8.00918C158.272 7.96406 161.038 7.94497 161.563 7.98489C161.552 8.02307 161.549 8.05951 161.533 8.09249C161.337 8.50033 161.135 8.90469 160.942 9.31253C160.833 9.54509 160.747 9.78632 160.638 10.0189C160.378 10.5656 160.109 11.1053 159.851 11.652C159.55 12.2941 159.259 12.9414 158.96 13.5853C158.651 14.25 158.338 14.9147 158.027 15.5794C157.731 16.2146 157.435 16.848 157.139 17.4832C156.802 18.2034 156.466 18.9236 156.131 19.6456C155.99 19.9476 155.854 20.253 155.717 20.5567C155.683 20.6296 155.639 20.6713 155.549 20.6713C155.107 20.6678 154.663 20.6713 154.203 20.6713L154.199 20.6661Z" fill="#2B4483"/> <path d="M144.361 20.668C144.201 20.6628 144.04 20.6593 143.88 20.6541C143.868 20.642 143.857 20.6298 143.845 20.6177C143.841 20.5656 143.836 20.5118 143.836 20.4597C143.836 16.3692 143.836 12.2804 143.836 8.18987V7.98682H147.007V20.668H144.446C144.416 20.668 144.388 20.668 144.358 20.668H144.361Z" fill="#2B4483"/> <path d="M145.613 6.2582C144.581 6.27034 144.005 5.89201 143.559 5.2655C143.293 4.8889 143.27 4.47065 143.286 4.04025C143.307 3.45018 143.614 2.99896 144.047 2.6293C144.324 2.39327 144.668 2.28047 145.031 2.20237C145.457 2.11039 145.847 2.21973 146.231 2.3603C146.517 2.46443 146.755 2.65533 146.956 2.88095C147.178 3.13085 147.34 3.41547 147.423 3.73827C147.648 4.60775 147.252 5.41822 146.608 5.87465C146.217 6.15233 145.865 6.2582 145.614 6.2582H145.613Z" fill="#2B4483"/> <path d="M98.2833 20.562V20.6679C98.2181 20.6332 98.2181 20.5967 98.2833 20.562Z" fill="#2B4483"/> <path d="M98.2833 20.562C98.2833 20.5967 98.2833 20.6332 98.2833 20.6679C98.2181 20.6332 98.2181 20.5967 98.2833 20.562Z" fill="#2B4483"/> <path d="M133.972 20.5532C133.972 20.5914 133.972 20.6296 133.972 20.6678C133.952 20.6452 133.931 20.6244 133.912 20.6018C133.931 20.5862 133.952 20.5688 133.972 20.5532Z" fill="#2B4483"/> <path d="M133.972 20.5532C133.972 20.5914 133.972 20.6296 133.972 20.6678C133.952 20.6452 133.931 20.6244 133.912 20.6018C133.931 20.5862 133.952 20.5688 133.972 20.5532Z" fill="#2B4483"/> <path d="M98.2833 20.562V20.6679C98.2181 20.6332 98.2181 20.5967 98.2833 20.562Z" fill="#2B4483"/> <path d="M98.2833 20.562C98.2833 20.5967 98.2833 20.6332 98.2833 20.6679C98.2181 20.6332 98.2181 20.5967 98.2833 20.562Z" fill="#2B4483"/> <path d="M98.2833 20.562C98.2833 20.5967 98.2833 20.6332 98.2833 20.6679C98.2181 20.6332 98.2181 20.5967 98.2833 20.562Z" fill="#2B4483"/> <path d="M135.219 20.5532C135.219 20.5914 135.219 20.6296 135.219 20.6678C135.2 20.6452 135.179 20.6244 135.159 20.6018C135.179 20.5862 135.2 20.5688 135.219 20.5532Z" fill="#2B4483"/> <path d="M98.2833 20.562C98.2833 20.5967 98.2833 20.6332 98.2833 20.6679C98.2181 20.6332 98.2181 20.5967 98.2833 20.562Z" fill="#2B4483"/> <path d="M135.219 20.5532C135.219 20.5914 135.219 20.6296 135.219 20.6678C135.2 20.6452 135.179 20.6244 135.159 20.6018C135.179 20.5862 135.2 20.5688 135.219 20.5532Z" fill="#2B4483"/> <path d="M98.2833 20.562C98.2833 20.5967 98.2833 20.6332 98.2833 20.6679C98.2181 20.6332 98.2181 20.5967 98.2833 20.562Z" fill="#2B4483"/> <path d="M83.5688 20.595C83.5688 20.6193 83.5723 20.6436 83.5723 20.6679C83.5195 20.647 83.5195 20.6227 83.5688 20.595Z" fill="#2B4483"/> <path d="M160.868 20.595C160.905 20.6175 160.916 20.6418 160.872 20.6679C160.817 20.6453 160.817 20.621 160.868 20.595Z" fill="#2B4483"/> <path d="M160.872 20.595C160.916 20.621 160.907 20.6453 160.868 20.6679C160.815 20.6418 160.817 20.6175 160.872 20.595Z" fill="#2B4483"/> <path d="M135.219 20.5898C135.219 20.6159 135.219 20.6419 135.219 20.6679C135.202 20.6523 135.184 20.6367 135.167 20.6211C135.184 20.6107 135.202 20.6003 135.219 20.5898Z" fill="#2B4483"/> <path d="M122.291 20.5898C122.291 20.6159 122.291 20.6419 122.29 20.6679C122.272 20.6575 122.254 20.6471 122.237 20.6367C122.254 20.6211 122.274 20.6055 122.291 20.5898Z" fill="#2B4483"/> <path d="M135.227 20.595C135.264 20.6175 135.276 20.6418 135.23 20.6679C135.175 20.647 135.174 20.6227 135.227 20.595Z" fill="#2B4483"/> <path d="M104.068 20.5898C104.068 20.6159 104.068 20.6419 104.068 20.6679C104.05 20.6523 104.033 20.6367 104.015 20.6211C104.033 20.6107 104.05 20.6003 104.068 20.5898Z" fill="#2B4483"/> <path d="M98.3115 20.5967C98.2974 20.6192 98.2833 20.6435 98.271 20.6661C98.2604 20.6505 98.2498 20.6349 98.2393 20.6192C98.2639 20.6123 98.2868 20.6036 98.3115 20.5967Z" fill="#2B4483"/> <path d="M144.264 20.6157C144.291 20.6157 144.317 20.6157 144.343 20.6157C144.328 20.6331 144.312 20.6504 144.296 20.6678C144.285 20.6504 144.275 20.6331 144.264 20.6157Z" fill="#2B4483"/> <path d="M138.122 20.6679C138.097 20.6662 138.073 20.6645 138.048 20.661C138.097 20.6176 138.108 20.6193 138.122 20.6679Z" fill="#2B4483"/> <path d="M104.068 20.6262V20.6679C104.054 20.6592 104.039 20.6505 104.025 20.6401C104.039 20.6349 104.054 20.6297 104.068 20.6262Z" fill="#2B4483"/> <path d="M138.17 20.6678C138.159 20.6556 138.147 20.6435 138.136 20.6331C138.147 20.6452 138.159 20.6574 138.17 20.6678Z" fill="#2B4483"/> <path d="M169.281 20.6678C169.283 20.6452 169.286 20.6226 169.288 20.6018C169.298 20.6174 169.311 20.6348 169.321 20.6504C169.307 20.6556 169.295 20.6625 169.281 20.6678Z" fill="#2B4483"/> <path d="M93.0472 20.6679C93.0331 20.6592 93.019 20.6505 93.0049 20.6401C93.019 20.6349 93.0331 20.6297 93.0472 20.6245C93.0472 20.6384 93.0472 20.6523 93.0472 20.6662V20.6679Z" fill="#2B4483"/> <path d="M114.502 7.52686C110.64 7.52686 108.046 10.1631 108.046 14.087C108.046 18.0109 110.7 20.6679 114.502 20.6679C117.357 20.6679 119.636 19.132 120.447 16.6572L120.521 16.4299H117.36L117.313 16.5271C116.775 17.6118 115.829 18.1619 114.502 18.1619C112.664 18.1619 111.352 17.0165 111.113 15.2255H120.681L120.704 15.0797C120.767 14.6823 120.797 14.2744 120.797 13.7937C120.797 10.104 118.208 7.52686 114.502 7.52686ZM111.146 12.8114C111.462 11.1158 112.716 10.0329 114.387 10.0329C116.247 10.0329 117.591 11.1662 117.718 12.8114H111.146Z" fill="#2B4483"/> <path d="M167.71 7.52686C163.849 7.52686 161.254 10.1631 161.254 14.087C161.254 18.0109 163.909 20.6679 167.71 20.6679C170.565 20.6679 172.844 19.132 173.655 16.6572L173.729 16.4299H170.569L170.521 16.5271C169.984 17.6118 169.037 18.1619 167.71 18.1619C165.872 18.1619 164.561 17.0165 164.321 15.2255H173.889L173.912 15.0797C173.976 14.6823 174.006 14.2744 174.006 13.7937C174.006 10.104 171.417 7.52686 167.71 7.52686ZM164.355 12.8114C164.67 11.1158 165.925 10.0329 167.596 10.0329C169.455 10.0329 170.8 11.1662 170.927 12.8114H164.355Z" fill="#2B4483"/> <path d="M141.786 10.3677V7.92933H138.727V4.85059H135.748V7.93107H134.264V10.3694H135.748V17.0545C135.748 19.4859 136.994 20.6678 139.555 20.6678H141.786V18.1843H140.012C139.039 18.1843 138.727 17.9101 138.727 17.0545V10.3694H141.786V10.3677Z" fill="#2B4483"/> <path d="M98.3716 20.6679H101.052V3.83545H98.3716V20.6679Z" fill="#2B4483"/> <path d="M103.735 20.6679H106.416V3.83545H103.735V20.6679Z" fill="#2B4483"/> </svg>`;
+export const menuIconUri = `data:image/svg+xml;utf8,<svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg"> <path fill-rule="evenodd" clip-rule="evenodd" d="M17.2915 5.83325C17.2915 6.17843 17.0117 6.45825 16.6665 6.45825L3.33317 6.45825C2.98799 6.45825 2.70817 6.17843 2.70817 5.83325C2.70817 5.48807 2.98799 5.20825 3.33317 5.20825L16.6665 5.20825C17.0117 5.20825 17.2915 5.48807 17.2915 5.83325Z" fill="#2B4483"/> <path fill-rule="evenodd" clip-rule="evenodd" d="M17.2915 10C17.2915 10.3452 17.0117 10.625 16.6665 10.625L3.33317 10.625C2.98799 10.625 2.70817 10.3452 2.70817 10C2.70817 9.65482 2.98799 9.375 3.33317 9.375L16.6665 9.375C17.0117 9.375 17.2915 9.65482 17.2915 10Z" fill="#2B4483"/> <path fill-rule="evenodd" clip-rule="evenodd" d="M17.2915 14.1667C17.2915 14.5119 17.0117 14.7917 16.6665 14.7917L3.33317 14.7917C2.98799 14.7917 2.70817 14.5119 2.70817 14.1667C2.70817 13.8216 2.98799 13.5417 3.33317 13.5417L16.6665 13.5417C17.0117 13.5417 17.2915 13.8216 17.2915 14.1667Z" fill="#2B4483"/> </svg>`;
+export const backIconUri = `data:image/svg+xml;utf8,<svg width="8" height="14" viewBox="0 0 8 14" fill="none" xmlns="http://www.w3.org/2000/svg"> <path fill-rule="evenodd" clip-rule="evenodd" d="M7.70711 0.292893C8.09763 0.683417 8.09763 1.31658 7.70711 1.70711L2.41421 7L7.70711 12.2929C8.09763 12.6834 8.09763 13.3166 7.70711 13.7071C7.31658 14.0976 6.68342 14.0976 6.29289 13.7071L0.292893 7.70711C-0.0976311 7.31658 -0.0976311 6.68342 0.292893 6.29289L6.29289 0.292893C6.68342 -0.0976311 7.31658 -0.0976311 7.70711 0.292893Z" fill="#2B4483"/> </svg>`;
+export const placeholderAvatarUri = `data:image/svg+xml;utf8,<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"> <circle cx="12" cy="12" r="12" fill="url(#pattern0)"/> <defs> <pattern id="pattern0" patternContentUnits="objectBoundingBox" width="1" height="1"> <use xlink:href="#image0_5405_761" transform="scale(0.015625)"/> </pattern> <image id="image0_5405_761" width="64" height="64" xlink:href="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAEAAAABACAYAAACqaXHeAAAAAXNSR0IArs4c6QAAAONJREFUeF7t20EOhEAIRFG4/6F7DvEnYeFzryQIv6pBd2behOu9dPvsbog+k+NLgArQAqmJcw9iAAhSgZKB3IJkkAySQTJ4CiE+gA8oBeg0mH3Ai084P89HhqwEqIA209ICsQdjAeaZIgaAYKxBDMCAYy8fXwAIgiAIcoJpJEYGI4VjB3YrbC9gL2AvkCB43cM5PgZgAAZgQFnNZAhdGykQBEEQBEEQDBmgAm2glM/z+QUYisYUGoldO7kY32IEAzCg6RgIRgjFAsw+AgRBMNYgBmCAT2TCYfoPPz/HCqQCX1eBHzHnv7C7WhBSAAAAAElFTkSuQmCC"/> </defs> </svg>`;
+export const logoutUri = `data:image/svg+xml;utf8,<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg"> <path d="M9.00195 7C9.01406 4.82497 9.11051 3.64706 9.87889 2.87868C10.7576 2 12.1718 2 15.0002 2L16.0002 2C18.8286 2 20.2429 2 21.1215 2.87868C22.0002 3.75736 22.0002 5.17157 22.0002 8L22.0002 16C22.0002 18.8284 22.0002 20.2426 21.1215 21.1213C20.2429 22 18.8286 22 16.0002 22H15.0002C12.1718 22 10.7576 22 9.87889 21.1213C9.11051 20.3529 9.01406 19.175 9.00195 17" stroke="#2B4483" stroke-width="1.5" stroke-linecap="round"/> <path d="M15 12L2 12M2 12L5.5 9M2 12L5.5 15" stroke="#2B4483" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"/> </svg>`;
+export const termsUri = `data:image/svg+xml;utf8,<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg"> <path d="M18 3.00001V3.00001C19.6569 3.00001 21 4.34315 21 6.00001L21 8.14286C21 8.47698 21 8.64405 20.9234 8.76602C20.8834 8.82962 20.8296 8.8834 20.766 8.92336C20.644 9 20.477 9 20.1429 9L15 9M18 3.00001V3.00001C16.3431 3.00001 15 4.34315 15 6.00001L15 9M18 3.00001L7 3.00001C5.11438 3.00001 4.17157 3.00001 3.58579 3.58579C3 4.17158 3 5.11439 3 7.00001L3 21L6 20L9 21L12 20L15 21L15 9" stroke="#2B4483" stroke-width="2"/> <path d="M7 7L11 7" stroke="#2B4483" stroke-width="2" stroke-linecap="round"/> <path d="M8 11H7" stroke="#2B4483" stroke-width="2" stroke-linecap="round"/> <path d="M7 15L10 15" stroke="#2B4483" stroke-width="2" stroke-linecap="round"/> </svg>`;
+export const privacyUri = `data:image/svg+xml;utf8,<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg"> <path fill-rule="evenodd" clip-rule="evenodd" d="M12 7.25C9.92893 7.25 8.25 8.92893 8.25 11C8.25 12.2277 8.84026 13.3168 9.75 14.0002V15C9.75 15.9665 10.5335 16.75 11.5 16.75H12.5C13.4665 16.75 14.25 15.9665 14.25 15V14.0002C15.1597 13.3168 15.75 12.2277 15.75 11C15.75 8.92893 14.0711 7.25 12 7.25ZM9.75 11C9.75 9.75736 10.7574 8.75 12 8.75C13.2426 8.75 14.25 9.75736 14.25 11C14.25 11.8319 13.7989 12.5593 13.1245 12.9494C12.8927 13.0835 12.75 13.3309 12.75 13.5987V15C12.75 15.1381 12.6381 15.25 12.5 15.25H11.5C11.3619 15.25 11.25 15.1381 11.25 15V13.5987C11.25 13.3309 11.1073 13.0835 10.8755 12.9494C10.2011 12.5593 9.75 11.8319 9.75 11Z" fill="#2B4483"/> <path fill-rule="evenodd" clip-rule="evenodd" d="M12 1.25C11.0625 1.25 10.1673 1.55658 8.72339 2.05112L7.99595 2.30014C6.51462 2.8072 5.3714 3.19852 4.55303 3.53099C4.14078 3.69846 3.78637 3.86067 3.50098 4.02641C3.22634 4.1859 2.95082 4.38484 2.76363 4.65154C2.5786 4.91516 2.48293 5.23924 2.42281 5.55122C2.36031 5.87556 2.32262 6.26464 2.2983 6.71136C2.25 7.59836 2.25 8.81351 2.25 10.3896V11.9914C2.25 18.0924 6.85803 21.0175 9.59833 22.2146L9.62543 22.2264C9.96523 22.3749 10.2846 22.5144 10.6516 22.6084C11.0391 22.7076 11.4507 22.75 12 22.75C12.5493 22.75 12.9609 22.7076 13.3484 22.6084C13.7154 22.5144 14.0348 22.3749 14.3745 22.2264L14.4017 22.2146C17.142 21.0175 21.75 18.0924 21.75 11.9914V10.3898C21.75 8.81361 21.75 7.5984 21.7017 6.71136C21.6774 6.26464 21.6397 5.87556 21.5772 5.55122C21.5171 5.23924 21.4214 4.91516 21.2364 4.65154C21.0492 4.38484 20.7737 4.1859 20.499 4.02641C20.2136 3.86067 19.8592 3.69846 19.447 3.53099C18.6286 3.19852 17.4854 2.8072 16.004 2.30013L15.2766 2.05112C13.8327 1.55658 12.9375 1.25 12 1.25ZM9.08062 3.5143C10.6951 2.96164 11.3423 2.75 12 2.75C12.6577 2.75 13.3049 2.96164 14.9194 3.5143L15.4922 3.71037C17.0048 4.22814 18.1079 4.60605 18.8824 4.92069C19.269 5.07774 19.5491 5.20935 19.7457 5.32353C19.8428 5.3799 19.9097 5.42642 19.9543 5.46273C19.9922 5.49349 20.0066 5.51092 20.0087 5.51348C20.0106 5.5166 20.0231 5.53737 20.0406 5.58654C20.0606 5.64265 20.0827 5.72309 20.1043 5.83506C20.148 6.06169 20.1811 6.37301 20.2039 6.79292C20.2497 7.63411 20.25 8.80833 20.25 10.4167V11.9914C20.25 17.1665 16.3801 19.7135 13.8012 20.84C13.4297 21.0023 13.2152 21.0941 12.9764 21.1552C12.7483 21.2136 12.47 21.25 12 21.25C11.53 21.25 11.2517 21.2136 11.0236 21.1552C10.7848 21.0941 10.5703 21.0023 10.1988 20.84C7.6199 19.7135 3.75 17.1665 3.75 11.9914V10.4167C3.75 8.80833 3.75028 7.63411 3.79608 6.79292C3.81894 6.37301 3.85204 6.06169 3.89571 5.83506C3.91729 5.72309 3.93944 5.64265 3.95943 5.58654C3.97693 5.5374 3.98936 5.51663 3.99129 5.51349C3.99336 5.51095 4.0078 5.49351 4.04567 5.46273C4.09034 5.42642 4.15722 5.3799 4.25429 5.32353C4.4509 5.20935 4.731 5.07774 5.11759 4.92069C5.8921 4.60605 6.99521 4.22814 8.5078 3.71037L9.08062 3.5143Z" fill="#2B4483"/> <path fill-rule="evenodd" clip-rule="evenodd" d="M12.0001 1.4001C11.0928 1.4001 10.2236 1.69598 8.77209 2.19312L8.04463 2.44215C6.56149 2.94983 5.4231 3.33956 4.60958 3.67005C4.19992 3.83648 3.85293 3.99563 3.57641 4.15622C3.3086 4.31175 3.05517 4.49751 2.88651 4.73781C2.71957 4.97566 2.62889 5.27514 2.5702 5.5797C2.50958 5.89428 2.47235 6.27558 2.44817 6.71961C2.40015 7.60154 2.4001 8.81176 2.4001 10.3897V11.9915C2.4001 17.9997 6.93405 20.8871 9.65848 22.0772L9.68558 22.089C10.0267 22.2381 10.3354 22.3727 10.6889 22.4632C11.0604 22.5583 11.4588 22.6001 12.0001 22.6001C12.5414 22.6001 12.9398 22.5583 13.3113 22.4632C13.6648 22.3727 13.9734 22.2381 14.3146 22.0891L14.3417 22.0772C17.0661 20.8871 21.6001 17.9997 21.6001 11.9915V10.3899C21.6001 8.81185 21.6 7.60158 21.552 6.71961C21.5278 6.27558 21.4906 5.89428 21.43 5.5797C21.3713 5.27514 21.2806 4.97566 21.1137 4.73781C20.945 4.49751 20.6916 4.31175 20.4238 4.15622C20.1473 3.99563 19.8003 3.83648 19.3906 3.67005C18.5771 3.33956 17.4387 2.94983 15.9556 2.44215L15.2281 2.19313C13.7766 1.69599 12.9074 1.4001 12.0001 1.4001ZM8.69807 1.90137C10.1211 1.41397 11.0376 1.1001 12.0001 1.1001C12.9626 1.1001 13.8791 1.41397 15.3021 1.90137L15.3253 1.9093L16.0578 2.16006C17.5349 2.66566 18.6812 3.05806 19.5035 3.39211C19.9183 3.56063 20.2802 3.7259 20.5744 3.89679C20.8559 4.06025 21.1535 4.27237 21.3592 4.56547C21.5624 4.85486 21.663 5.20355 21.7246 5.52293C21.789 5.85704 21.8271 6.25388 21.8516 6.7033C21.9001 7.59443 21.9001 8.81286 21.9001 10.3847V11.9915C21.9001 18.1853 17.218 21.1481 14.4618 22.3521C14.4618 22.3521 14.4618 22.3521 14.4618 22.3521L14.43 22.366C14.0932 22.5132 13.7644 22.6568 13.3857 22.7538C12.9823 22.8571 12.5573 22.9001 12.0001 22.9001C11.4429 22.9001 11.0179 22.8571 10.6145 22.7538C10.2357 22.6568 9.90695 22.5132 9.57015 22.366L9.53839 22.3521C6.7822 21.1481 2.1001 18.1853 2.1001 11.9915V10.3844C2.1001 8.81273 2.1001 7.59438 2.14862 6.7033C2.17309 6.25388 2.21124 5.85704 2.27562 5.52293C2.33717 5.20355 2.43783 4.85486 2.64095 4.56547C2.84667 4.27237 3.14429 4.06025 3.42575 3.89679C3.72 3.7259 4.08185 3.56063 4.49667 3.39211C5.31895 3.05807 6.46528 2.66567 7.94232 2.16008L7.94746 2.15832L8.67488 1.90931L8.69807 1.90137ZM9.03214 3.37248C10.6367 2.82322 11.3115 2.6001 12.0001 2.6001C12.6887 2.6001 13.3635 2.82322 14.9681 3.37248L15.5409 3.56855C17.0516 4.08568 18.1596 4.4652 18.939 4.78181C19.3281 4.9399 19.6156 5.07457 19.8211 5.19392C19.9229 5.253 19.9967 5.30387 20.049 5.34643C20.0942 5.38311 20.116 5.40771 20.1251 5.41885C20.1293 5.42395 20.1331 5.42932 20.1365 5.43492C20.1446 5.44797 20.1614 5.4784 20.182 5.53631C20.2052 5.60149 20.2291 5.68966 20.2517 5.80677C20.2972 6.04319 20.3308 6.36229 20.3538 6.78486C20.3999 7.63114 20.4001 8.81035 20.4001 10.4168V11.9915C20.4001 17.2589 16.4567 19.8438 13.8614 20.9775C13.4922 21.1388 13.2664 21.236 13.0137 21.3007C12.7697 21.3631 12.478 21.4001 12.0001 21.4001C11.5222 21.4001 11.2305 21.3631 10.9865 21.3007C10.7338 21.236 10.508 21.1388 10.1388 20.9775C7.5435 19.8438 3.6001 17.2589 3.6001 11.9915V10.4168C3.6001 8.81035 3.60032 7.63114 3.6464 6.78486C3.66941 6.36229 3.70296 6.04319 3.74852 5.80677C3.77109 5.68966 3.79501 5.60149 3.81822 5.5363C3.83883 5.47844 3.85561 5.44802 3.86364 5.43497C3.8671 5.42935 3.87092 5.42397 3.87508 5.41886C3.88413 5.40775 3.90597 5.38315 3.95115 5.34643C4.00351 5.30387 4.07731 5.253 4.17905 5.19392C4.38455 5.07457 4.6721 4.9399 5.06123 4.78181C5.84061 4.4652 6.9486 4.08568 8.45932 3.56855L9.03214 3.37248ZM12.0001 2.9001C11.3732 2.9001 10.7538 3.10025 9.12929 3.65631L8.55648 3.85239C7.04201 4.37079 5.94378 4.74709 5.17415 5.05975C4.79011 5.21577 4.51744 5.34432 4.32972 5.45334C4.23733 5.50699 4.17736 5.54916 4.14038 5.57922C4.12864 5.58877 4.12036 5.59616 4.11491 5.60131C4.11154 5.60882 4.10672 5.62044 4.10083 5.63696C4.08408 5.684 4.06369 5.75671 4.0431 5.86354C4.00131 6.08039 3.96868 6.38393 3.94596 6.80117C3.90043 7.63727 3.9001 8.80651 3.9001 10.4168V11.9915C3.9001 17.0742 7.69648 19.5833 10.2589 20.7026C10.6328 20.8659 10.8359 20.9524 11.0609 21.01C11.273 21.0643 11.538 21.1001 12.0001 21.1001C12.4622 21.1001 12.7272 21.0643 12.9393 21.01C13.1642 20.9524 13.3674 20.8659 13.7413 20.7026C16.3037 19.5833 20.1001 17.0742 20.1001 11.9915V10.4168C20.1001 8.80651 20.0998 7.63727 20.0542 6.80117C20.0315 6.38393 19.9989 6.08039 19.9571 5.86354C19.9365 5.75672 19.9161 5.684 19.8994 5.63696C19.8935 5.62045 19.8886 5.60882 19.8853 5.60131C19.8798 5.59616 19.8716 5.58877 19.8598 5.57922C19.8228 5.54917 19.7629 5.50699 19.6705 5.45334C19.4828 5.34432 19.2101 5.21577 18.826 5.05975C18.0564 4.74709 16.9582 4.37079 15.4437 3.85239L14.8709 3.65631C13.2464 3.10025 12.627 2.9001 12.0001 2.9001ZM12.0001 7.4001C10.0119 7.4001 8.4001 9.01187 8.4001 11.0001C8.4001 12.1786 8.9665 13.2241 9.84018 13.8803C9.8779 13.9087 9.9001 13.9531 9.9001 14.0003V15.0001C9.9001 15.8838 10.6164 16.6001 11.5001 16.6001H12.5001C13.3838 16.6001 14.1001 15.8838 14.1001 15.0001V14.0003C14.1001 13.9531 14.1223 13.9087 14.16 13.8803C15.0337 13.2241 15.6001 12.1786 15.6001 11.0001C15.6001 9.01187 13.9883 7.4001 12.0001 7.4001ZM8.1001 11.0001C8.1001 8.84619 9.84619 7.1001 12.0001 7.1001C14.154 7.1001 15.9001 8.84619 15.9001 11.0001C15.9001 12.2498 15.3118 13.3615 14.4001 14.0743V15.0001C14.4001 16.0494 13.5494 16.9001 12.5001 16.9001H11.5001C10.4508 16.9001 9.6001 16.0494 9.6001 15.0001V14.0743C8.68837 13.3615 8.1001 12.2498 8.1001 11.0001ZM9.6001 11.0001C9.6001 9.67461 10.6746 8.6001 12.0001 8.6001C13.3256 8.6001 14.4001 9.67461 14.4001 11.0001C14.4001 11.8876 13.9187 12.6635 13.1997 13.0794C13.0143 13.1866 12.9001 13.3846 12.9001 13.5988V15.0001C12.9001 15.221 12.721 15.4001 12.5001 15.4001H11.5001C11.2792 15.4001 11.1001 15.221 11.1001 15.0001V13.5988C11.1001 13.3846 10.9859 13.1866 10.8005 13.0794C10.0815 12.6635 9.6001 11.8876 9.6001 11.0001ZM12.0001 8.9001C10.8403 8.9001 9.9001 9.8403 9.9001 11.0001C9.9001 11.7763 10.3208 12.4553 10.9508 12.8197C11.2288 12.9806 11.4001 13.2775 11.4001 13.5988V15.0001C11.4001 15.0553 11.4449 15.1001 11.5001 15.1001H12.5001C12.5553 15.1001 12.6001 15.0553 12.6001 15.0001V13.5988C12.6001 13.2775 12.7714 12.9806 13.0494 12.8197C13.6794 12.4553 14.1001 11.7763 14.1001 11.0001C14.1001 9.8403 13.1599 8.9001 12.0001 8.9001Z" fill="#2B4483"/> </svg>`;
+export const reactUri = `data:image/svg+xml;utf8,<svg width="22" height="22" viewBox="0 0 22 22" fill="none" xmlns="http://www.w3.org/2000/svg"> <path fill-rule="evenodd" clip-rule="evenodd" d="M12.4618 3.04683C13.6652 3.86925 14.8624 4.85943 16.0015 5.99857C17.1407 7.13773 18.1309 8.33494 18.9533 9.53839C19.6544 8.16096 20.0758 6.85665 20.2064 5.72508C20.371 4.298 20.0678 3.24358 19.4122 2.58796C18.7566 1.93235 17.7022 1.62912 16.2751 1.79377C15.1435 1.92433 13.8392 2.34576 12.4618 3.04683ZM19.8687 11.0001C20.8741 9.23497 21.5142 7.47673 21.6965 5.89701C21.8915 4.20664 21.5677 2.62215 20.4729 1.5273C19.378 0.432462 17.7935 0.108624 16.1032 0.303658C14.5235 0.485924 12.7652 1.12603 11.0001 2.13142C9.23497 1.12609 7.47679 0.486026 5.89713 0.303785C4.2068 0.108777 2.62236 0.432625 1.52754 1.52744C0.432717 2.62227 0.108875 4.20674 0.303895 5.89709C0.486149 7.47678 1.12623 9.23499 2.13159 11.0001C1.12632 12.7652 0.486293 14.5233 0.304075 16.103C0.109092 17.7933 0.432947 19.3777 1.52774 20.4725C2.62254 21.5673 4.20694 21.8911 5.89723 21.6961C7.47686 21.5139 9.235 20.8739 11.0001 19.8686C12.7652 20.8739 14.5234 21.514 16.1031 21.6963C17.7934 21.8913 19.3778 21.5674 20.4727 20.4726C21.5675 19.3778 21.8913 17.7934 21.6963 16.103C21.5141 14.5234 20.874 12.7652 19.8687 11.0001ZM18.1195 11.0001C17.2825 9.67499 16.2143 8.33261 14.9409 7.05923C13.6675 5.78587 12.3252 4.71761 11.0001 3.88069C9.67508 4.71759 8.3328 5.7858 7.05951 7.05909C5.78609 8.33252 4.71778 9.67495 3.88085 11.0001C4.71776 12.3252 5.78599 13.6675 7.05932 14.9408C8.33267 16.2142 9.67501 17.2824 11.0001 18.1193C12.3252 17.2824 13.6677 16.2141 14.9411 14.9407C16.2144 13.6674 17.2826 12.3251 18.1195 11.0001ZM12.4618 18.9532C13.6653 18.1307 14.8625 17.1405 16.0017 16.0013C17.1408 14.8623 18.1309 13.6652 18.9533 12.4618C19.6543 13.8392 20.0757 15.1435 20.2062 16.275C20.3708 17.702 20.0676 18.7563 19.412 19.4119C18.7564 20.0675 17.702 20.3708 16.275 20.2061C15.1435 20.0756 13.8392 19.6542 12.4618 18.9532ZM9.53833 18.9532C8.33492 18.1308 7.13777 17.1406 5.99866 16.0015C4.85956 14.8624 3.86943 13.6653 3.04702 12.4619C2.34607 13.8392 1.92471 15.1434 1.79419 16.2749C1.62958 17.7019 1.93281 18.7562 2.5884 19.4118C3.24399 20.0674 4.29835 20.3706 5.72534 20.206C6.85681 20.0755 8.161 19.6541 9.53833 18.9532ZM3.04699 9.53837C3.86944 8.33488 4.85966 7.13763 5.99885 5.99843C7.13792 4.85936 8.33502 3.86924 9.53837 3.04685C8.16098 2.34583 6.85673 1.92444 5.72522 1.7939C4.29819 1.62927 3.2438 1.9325 2.58821 2.5881C1.9326 3.24371 1.62937 4.29811 1.79401 5.72517C1.92456 6.8567 2.34596 8.16097 3.04699 9.53837ZM11.0002 9.24995C10.0337 9.24995 9.2502 10.0335 9.2502 11C9.2502 11.9665 10.0337 12.75 11.0002 12.75C11.9667 12.75 12.7502 11.9665 12.7502 11C12.7502 10.0335 11.9667 9.24995 11.0002 9.24995ZM7.7502 11C7.7502 9.20503 9.20527 7.74995 11.0002 7.74995C12.7951 7.74995 14.2502 9.20503 14.2502 11C14.2502 12.7949 12.7951 14.25 11.0002 14.25C9.20527 14.25 7.7502 12.7949 7.7502 11Z" fill="#2B4483"/> </svg>`;
+export const closeUri = `data:image/svg+xml;utf8,<svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg"> <path d="M15 5L5 15" stroke="#2B4483" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/> <path d="M5 5L15 15" stroke="#2B4483" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/> </svg>`;
+export const metamaskLogoUri = `data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" xml:space="preserve" id="Layer_1" x="0" y="0" viewBox="0 0 318.6 318.6"><style>.st1,.st6{fill:#e4761b;stroke:#e4761b;stroke-linecap:round;stroke-linejoin:round}.st6{fill:#f6851b;stroke:#f6851b}</style><path fill="#e2761b" stroke="#e2761b" stroke-linecap="round" stroke-linejoin="round" d="m274.1 35.5-99.5 73.9L193 65.8z"/><path d="m44.4 35.5 98.7 74.6-17.5-44.3zm193.9 171.3-26.5 40.6 56.7 15.6 16.3-55.3zm-204.4.9L50.1 263l56.7-15.6-26.5-40.6z" class="st1"/><path d="m103.6 138.2-15.8 23.9 56.3 2.5-2-60.5zm111.3 0-39-34.8-1.3 61.2 56.2-2.5zM106.8 247.4l33.8-16.5-29.2-22.8zm71.1-16.5 33.9 16.5-4.7-39.3z" class="st1"/><path fill="#d7c1b3" stroke="#d7c1b3" stroke-linecap="round" stroke-linejoin="round" d="m211.8 247.4-33.9-16.5 2.7 22.1-.3 9.3zm-105 0 31.5 14.9-.2-9.3 2.5-22.1z"/><path fill="#233447" stroke="#233447" stroke-linecap="round" stroke-linejoin="round" d="m138.8 193.5-28.2-8.3 19.9-9.1zm40.9 0 8.3-17.4 20 9.1z"/><path fill="#cd6116" stroke="#cd6116" stroke-linecap="round" stroke-linejoin="round" d="m106.8 247.4 4.8-40.6-31.3.9zM207 206.8l4.8 40.6 26.5-39.7zm23.8-44.7-56.2 2.5 5.2 28.9 8.3-17.4 20 9.1zm-120.2 23.1 20-9.1 8.2 17.4 5.3-28.9-56.3-2.5z"/><path fill="#e4751f" stroke="#e4751f" stroke-linecap="round" stroke-linejoin="round" d="m87.8 162.1 23.6 46-.8-22.9zm120.3 23.1-1 22.9 23.7-46zm-64-20.6-5.3 28.9 6.6 34.1 1.5-44.9zm30.5 0-2.7 18 1.2 45 6.7-34.1z"/><path d="m179.8 193.5-6.7 34.1 4.8 3.3 29.2-22.8 1-22.9zm-69.2-8.3.8 22.9 29.2 22.8 4.8-3.3-6.6-34.1z" class="st6"/><path fill="#c0ad9e" stroke="#c0ad9e" stroke-linecap="round" stroke-linejoin="round" d="m180.3 262.3.3-9.3-2.5-2.2h-37.7l-2.3 2.2.2 9.3-31.5-14.9 11 9 22.3 15.5h38.3l22.4-15.5 11-9z"/><path fill="#161616" stroke="#161616" stroke-linecap="round" stroke-linejoin="round" d="m177.9 230.9-4.8-3.3h-27.7l-4.8 3.3-2.5 22.1 2.3-2.2h37.7l2.5 2.2z"/><path fill="#763d16" stroke="#763d16" stroke-linecap="round" stroke-linejoin="round" d="m278.3 114.2 8.5-40.8-12.7-37.9-96.2 71.4 37 31.3 52.3 15.3 11.6-13.5-5-3.6 8-7.3-6.2-4.8 8-6.1zM31.8 73.4l8.5 40.8-5.4 4 8 6.1-6.1 4.8 8 7.3-5 3.6 11.5 13.5 52.3-15.3 37-31.3-96.2-71.4z"/><path d="m267.2 153.5-52.3-15.3 15.9 23.9-23.7 46 31.2-.4h46.5zm-163.6-15.3-52.3 15.3-17.4 54.2h46.4l31.1.4-23.6-46zm71 26.4 3.3-57.7 15.2-41.1h-67.5l15 41.1 3.5 57.7 1.2 18.2.1 44.8h27.7l.2-44.8z" class="st6"/></svg>`;
+export const walletConnectLogoUri = `data:image/svg+xml;utf8,<svg fill="none" height="332" viewBox="0 0 480 332" width="480" xmlns="http://www.w3.org/2000/svg"><path d="m126.613 93.9842c62.622-61.3123 164.152-61.3123 226.775 0l7.536 7.3788c3.131 3.066 3.131 8.036 0 11.102l-25.781 25.242c-1.566 1.533-4.104 1.533-5.67 0l-10.371-10.154c-43.687-42.7734-114.517-42.7734-158.204 0l-11.107 10.874c-1.565 1.533-4.103 1.533-5.669 0l-25.781-25.242c-3.132-3.066-3.132-8.036 0-11.102zm280.093 52.2038 22.946 22.465c3.131 3.066 3.131 8.036 0 11.102l-103.463 101.301c-3.131 3.065-8.208 3.065-11.339 0l-73.432-71.896c-.783-.767-2.052-.767-2.835 0l-73.43 71.896c-3.131 3.065-8.208 3.065-11.339 0l-103.4657-101.302c-3.1311-3.066-3.1311-8.036 0-11.102l22.9456-22.466c3.1311-3.065 8.2077-3.065 11.3388 0l73.4333 71.897c.782.767 2.051.767 2.834 0l73.429-71.897c3.131-3.065 8.208-3.065 11.339 0l73.433 71.897c.783.767 2.052.767 2.835 0l73.431-71.895c3.132-3.066 8.208-3.066 11.339 0z" fill="#3396ff"/></svg>`;
